requirements.txt
pillow
moviepy
fastapi
uvicorn
python-dotenv